REP-196 update tests
